# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.WalletRequest:
    properties:
      amount:
        type: integer
      operationType:
        type: string
      walletId:
        type: string
    required:
    - amount
    - operationType
    - walletId
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления операциями кошелька
  title: Wallet Service API
  version: "1.0"
paths:
  /wallet:
    post:
      consumes:
      - application/json
      description: Пополнение или списание средств с кошелька.
      parameters:
      - description: Запрос для обновления кошелька
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.WalletRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка на стороне сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить баланс кошелька
      tags:
      - Wallet
  /wallets:
    post:
      consumes:
      - application/json
      description: Создает новый кошелек с начальным балансом. Баланс должен быть
        неотрицательным.
      parameters:
      - description: Начальный баланс (неотрицательное число)
        in: body
        name: input
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "201":
          description: ID созданного кошелька
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный запрос, например, отрицательный баланс
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка на стороне сервера, например, проблема с базой данных
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать кошелек
      tags:
      - Wallet
  /wallets/{walletId}:
    get:
      consumes:
      - application/json
      description: Возвращает текущий баланс кошелька по его UUID.
      parameters:
      - description: UUID кошелька
        in: path
        name: walletId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Баланс кошелька
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный UUID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Кошелек не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить баланс кошелька
      tags:
      - Wallet
swagger: "2.0"